+++ bingads/v10/bulk/entities/ad_extensions/__init__.py
@@ -8,3 +8,4 @@ from .bulk_app_ad_extensions import *
 from .bulk_callout_ad_extensions import *
 from .bulk_review_ad_extensions import *
 from .bulk_structured_snippet_ad_extensions import *
+from .bulk_sitelink2_ad_extensions import *